requirements.txt
Django==4.2.7
pymysql==1.1.0
crispy-bootstrap5==0.7
Pillow==10.1.0
djangorestframework==3.14.0
django-crispy-forms==2.0
python-dotenv==1.0.0
qrcode==7.4.2
django-debug-toolbar==4.3.0